feat: add mjs solution
